Lesson 4. Task 5.
