+++ Lesson4/generated_file_with_array.txt
@@ -1,0 +1,8 @@
+1
+2
+3
+4
+5
+6
+7
+8
